First test for backtrace-read + symbol resolution (killed the write-profile-to-disk stuff because the test on java side is long gone, we have much better tests that actually send profiles across over http-calls now, we it was no longer needed)
