+++ recorder/CMakeLists.txt
@@ -58,7 +58,7 @@ set(OUTPUT "fkpagent")
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${BIN})
 #protobuf idl setup
 add_subdirectory(src/idl)
-include_directories(src/idl)
+include_directories(src/idl src/main/cpp)
 
 ##########################################################
 # Libraries & Code
@@ -125,6 +125,7 @@ set(SOURCE_FILES
     ${SRC}/profiler.cc
     ${SRC}/profiler.hh
     ${SRC}/stacktraces.hh
+    ${SRC}/stacktraces.cc
     ${SRC}/trace.hh
     ${SRC}/thread_map.hh
     ${SRC}/thread_map.cc
@@ -146,9 +147,7 @@ set(SOURCE_FILES
     ${SRC}/prob_pct.cc
     ${SRC}/metric_formatter.hh
     ${SRC}/metric_formatter.cc
-    ${SRC}/unique_readsafe_ptr.hh
-    ${SRC}/backtracer.hh
-    ${SRC}/backtracer.cc)
+    ${SRC}/unique_readsafe_ptr.hh)
 
 set(TI_SRC_FILES
     ${SRC}/agent.cc)
@@ -171,11 +170,12 @@ set(TEST_FILES
     ${SRC_TEST}/test_blocking_ring_buffer.cc
     ${SRC_TEST}/test_syslog_tsdb_metric_formatter.cc
     ${SRC_TEST}/test_unique_readsafe_ptr.cc
+    ${SRC_TEST}/test_site_resolver.cc
     ${SRC_TEST}/test.cc)
 
 set(TEST_UTIL_FILES
-    ${SRC_TEST}/test_profile.hh
-    ${SRC_TEST}/test_profile.cc)
+    ${SRC_TEST}/test_helpers.hh
+    ${SRC_TEST}/test_helpers.cc)
 
 set(TEST_JNI_FILES
     ${SRC_TEST}/test_jni.hh
@@ -221,7 +221,7 @@ if (${VERBOSE_LOGS} MATCHES "y")
   set(CMAKE_CXX_FLAGS_DEBUG  "${CMAKE_CXX_FLAGS_DEBUG} -DSPDLOG_TRACE_ON -DSPDLOG_DEBUG_ON")
 endif()
 # to enable tracing add: -DENABLE_TRACING
-set(CMAKE_CXX_FLAGS "-I${INCLUDE_VENDOR_SPDLOG} -I${INCLUDE_VENDOR_CONCQUEUE} -I${PROTOBUF_INCLUDE_DIR} ${CMAKE_CXX_FLAGS} ${GLOBAL_WARNINGS} ${GLOBAL_COPTS} -pthread -std=c++0x")
+set(CMAKE_CXX_FLAGS "-I${INCLUDE_VENDOR_SPDLOG} -I${INCLUDE_VENDOR_CONCQUEUE} -I${PROTOBUF_INCLUDE_DIR} ${CMAKE_CXX_FLAGS} ${GLOBAL_WARNINGS} ${GLOBAL_COPTS} -pthread -std=c++0x -fno-omit-frame-pointer")
 
 ##########################################################
 # Build
@@ -238,21 +238,21 @@ target_link_libraries(${OUTPUT} ${BASE_OUTPUT})
 
 add_library(testutil SHARED ${TEST_UTIL_FILES})
 add_library(testbase STATIC ${TEST_BASE_FILES})
-add_executable(unitTests ${TEST_FILES} ${SRC_TEST}/main.cc)
+add_executable(tests ${TEST_FILES} ${SRC_TEST}/main.cc)
 add_library(testjni SHARED ${TEST_JNI_FILES})
 
 if (DEFINED ENV{UNITTEST_LIBRARIES})
   message("User has configured " $ENV{UNITTEST_LIBRARIES} " as the unit test libraries")
-  target_link_libraries(unitTests testutil testbase ${BASE_OUTPUT} $ENV{UNITTEST_LIBRARIES})
+  target_link_libraries(tests testutil testbase ${BASE_OUTPUT} $ENV{UNITTEST_LIBRARIES})
   target_link_libraries(testjni testutil testbase ${BASE_OUTPUT} $ENV{UNITTEST_LIBRARIES})
 else()
-  target_link_libraries(unitTests testutil testbase ${BASE_OUTPUT} ${unittest++_LIBRARIES})
+  target_link_libraries(tests testutil testbase ${BASE_OUTPUT} ${unittest++_LIBRARIES})
   target_link_libraries(testjni testutil testbase ${BASE_OUTPUT} ${unittest++_LIBRARIES})
 endif()
 
 
 # make test
-add_test(unitTests ${BIN}/unitTests)
+add_test(tests ${BIN}/tests)
 
 # install
 install(TARGETS  ${OUTPUT}
