+++ recorder/src/main/cpp/profile_writer.hh
@@ -10,7 +10,6 @@
 #include "circular_queue.hh"
 #include <unordered_map>
 #include <unordered_set>
-#include "backtracer.hh"
 
 #define NOCTX_NAME "~ OTHERS ~"
 
@@ -110,7 +109,7 @@ private:
     metrics::Mtr& s_m_stack_sample_err;
     metrics::Mtr& s_m_cpu_sample_add;
 
-    Backtracer::SymInfo syms;
+    SiteResolver::SymInfo syms;
 
     CtxId report_ctx(PerfCtx::TracePt trace_pt);
