+++ recorder/src/main/cpp/profiler.hh
@@ -100,7 +100,7 @@ private:
     metrics::Hist& s_h_pop_spree_len;
     metrics::Timer& s_t_pop_spree_tm;
 
-    Backtracer::SymInfo si;
+    SiteResolver::SymInfo si;
 
     void set_sampling_freq(std::uint32_t sampling_freq);
