+++ recorder/src/test/cpp/test_helpers.cc
@@ -1,0 +1,7 @@
+#include "test_helpers.hh"
+#include <unordered_map>
+#include <fstream>
+
+__attribute__ ((noinline)) void some_λ_caller(std::function<void()> fn) {
+    fn();
+}
