--- recorder/src/test/cpp/test_profile.hh
@@ -1,8 +1,0 @@
-#ifndef TEST_PROFILE_H
-#define TEST_PROFILE_H
-
-#include <string>
-
-void generate_cpusample_simple_profile(const std::string& profile_data_file);
-
-#endif /* TEST_PROFILE_H */
